Incorrect MPI datatype used for a MPI_Bcast call. Fixed now.
git-svn-id: svn+ssh://svn.wcas.northwestern.edu/var/svn/cmc/branches/cmc-mpi@1011 46abec54-52ec-0310-8d7b-a975267b85f2

+++ cmc_utils.c
@@ -1536,7 +1536,7 @@ void central_calculate(void)
 	buf_bcast_long[1] = central.N_sin;
 	buf_bcast_long[2] = central.N_bin;
 
-	MPI_Bcast( buf_bcast_long, 3, MPI_DOUBLE, 0, MPI_COMM_WORLD );
+	MPI_Bcast( buf_bcast_long, 3, MPI_LONG, 0, MPI_COMM_WORLD );
 
 	Ncentral = buf_bcast_long[0];
 	central.N_sin = buf_bcast_long[1];
